+++ src/AzureAIProxy.Management/Components/Pages/ModelList.razor.cs
@@ -29,7 +29,7 @@ public partial class ModelList : ComponentBase
         var dialog = await DialogService.ShowAsync<DeleteConfirmation>("Delete Record", parameters, options);
         var result = await dialog.Result;
 
-        if (!result.Canceled)
+        if (result is not null && !result.Canceled)
         {
             await ModelService.DeleteOwnerCatalogAsync(resource.CatalogId);
             Models = await ModelService.GetOwnerCatalogsAsync();
